kompilieren/backend/pyproject.toml
[project]
name = "landly-backend"
version = "1.0.0"
description = "Landly Backend – FastAPI Server"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.44",
    "geoalchemy2>=0.17.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.2.1",
    "pydantic[email]>=2.12.5",
]

[project.scripts]
start = "run_backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
